requirements.txt
streamlit==1.27.0
pandas==2.1.0
python-dotenv==1.0.1
openai==1.29.0
supabase==1.1.0
